+++ src/kernel/text_mode/text_mode.h
@@ -1,13 +1,13 @@
 #include "kernel/string/string.h"
 namespace kernel {
   namespace text_mode {
-    inline int put_char (char character, unsigned short int line,
-                         unsigned short int column);
+    inline int put_char (char character, unsigned int line,
+                         unsigned int column);
     // int puts(const tacospp::kernel::string::String *string,
     //          unsigned short int line, unsigned short int column);
 
     int puts(const tacospp::kernel::string::String &string,
-             unsigned short int line, unsigned short int column);
+             unsigned int line, unsigned int column);
 
     /** Put a C string to console.
 
@@ -15,9 +15,9 @@ namespace kernel {
         console easier as postfix literal prefixes are not yet in gcc
         (specified in draft C++0x standard).
      */
-    int puts(const char* string, unsigned short int line,
-             unsigned short int column);
-    int put_hex(unsigned int number,unsigned short int line, unsigned short int column);
+    int puts(const char* string, unsigned int line,
+             unsigned int column);
+    int put_hex(unsigned int number,unsigned int line, unsigned int column);
     char hex2char(unsigned int hex_number);
 
     /** Clear a line on the console.
